Refactor UIManager and cleanup utils
