+++ src/utils/IO_utility.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <string>
+#include <cstdio>
 
 namespace IO_utility {
 
@@ -8,18 +9,10 @@ public:
     std::string ReadTextFile(const std::string &path) const;
 };
 
-class FileDeleter {
-public:
-    void DeleteFile(const std::string &path) const;
-};
-
 class FileIO {
 public:
     static std::string ReadTextFile(const std::string &path) {
         return FileReader{}.ReadTextFile(path);
-    }
-    static void DeleteFile(const std::string &path) {
-        FileDeleter{}.DeleteFile(path);
     }
 };
 
@@ -28,7 +21,7 @@ inline std::string readFile(const char *path) {
 }
 
 inline void deleteFile(const char *path) {
-    FileIO::DeleteFile(path);
+    std::remove(path);
 }
 
 } // namespace IO_utility
